test: fix all tests except cif_ensemble
