+++ .gitignore
@@ -65,6 +65,10 @@ coverage.xml
 
 # Sphinx documentation
 docs/build/
+tests/data
+AXUGAE.cif
+FIXVUH.cif
+NUMSIB.cif
 docs/source/generated/
 
 # pytest
