+++ news/return-obj-valkue.rst
@@ -1,5 +1,6 @@
 **Added:**
 
+* Implement compute_CN to separate computing CN related metrics, separated from computing connections.
 * Return objective function value after finding refined CIF radius.
 
 **Changed:**
