+++ tests/test_distances.py
@@ -11,4 +11,5 @@ def test_expected_shortest_distance(cif_CUMNON_sb: Cif):
         (("Sb", "Sb"), 7.813),
         (("Sb", "Sb"), 7.88),
     ]
+    cif_CUMNON_sb.compute_CN()
     assert cif_CUMNON_sb.CN_max_by_min_dist_method == 8
